+++ packages/terra-application-name/src/terra-dev-site/doc/example/ApplicationMenuNameStandard.jsx
@@ -6,17 +6,17 @@ import Placeholder from 'terra-application-name/lib/terra-dev-site/doc/common/Pl
 import ApplicationMenuName from 'terra-application-name/lib/ApplicationMenuName';
 /* eslint-enable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 
-const size = { height: '450px', width: '300px' };
+const demosize = { height: '450px', width: '300px' };
+const demobgcolor = { backgroundColor: 'rgb(140, 161, 177)' };
 const ApplicationMenuNameStandard = () => (
   <Menu
-    style={size}
+    style={demosize}
     header={(
-      <div style={{ backgroundColor: 'green', width: '100%' }}>
-        <ApplicationMenuName
-          title="App-Name"
-          accessory={<Image variant="rounded" src="https://github.com/cerner/terra-framework/raw/master/terra.png" height="26px" width="26px" isFluid />}
-        />
-      </div>
+      <ApplicationMenuName
+        title="App-Name"
+        accessory={<Image variant="rounded" src="https://github.com/cerner/terra-framework/raw/master/terra.png" height="26px" width="26px" isFluid />}
+        style={demobgcolor}
+      />
     )}
     extensions={<Placeholder text="Extensions" height="50px" />}
     content={<Placeholder text="Content" />}