+++ packages/terra-popup/src/_PopupContent.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { injectIntl, intlShape } from 'react-intl';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import Button from 'terra-button';
@@ -51,6 +52,10 @@ const propTypes = {
    */
   contentWidthMax: PropTypes.number,
   /**
+   * @private The intl object to be injected for translations. Provided by the injectIntl function.
+   */
+  intl: intlShape.isRequired,
+  /**
    * Set this to `true` if your content has focusable elements and you want them to receive focus instead of focusing on the default popup frame when the popup is opened.
    */
   isFocusedDisabled: PropTypes.bool,
@@ -95,9 +100,9 @@ const defaultProps = {
 };
 
 class PopupContent extends React.Component {
-  static addPopupHeader(children, onRequestClose) {
+  static addPopupHeader(children, onRequestClose, close) {
     const icon = <span className={cx('close-icon')} />;
-    const button = <Button variant="utility" className={cx('close')} isIconOnly icon={icon} onClick={onRequestClose} />;
+    const button = <Button variant="utility" isIconOnly icon={icon} onClick={onRequestClose} text={close} />;
     const header = <div className={cx('header')}>{button}</div>;
     return <ContentContainer header={header} fill>{children}</ContentContainer>;
   }
@@ -178,6 +183,7 @@ class PopupContent extends React.Component {
       contentHeightMax,
       contentWidth,
       contentWidthMax,
+      intl,
       isFocusedDisabled,
       isHeaderDisabled,
       isHeightAutomatic,
@@ -198,7 +204,8 @@ class PopupContent extends React.Component {
 
     let content = PopupContent.cloneChildren(children, isHeightAutomatic, isWidthAutomatic, isHeightBounded, isWidthBounded, isHeaderDisabled);
     if (isFullScreen && !isHeaderDisabled) {
-      content = PopupContent.addPopupHeader(content, onRequestClose);
+      const close = intl.formatMessage({ id: 'Terra.popup.header.close' });
+      content = PopupContent.addPopupHeader(content, onRequestClose, close);
     }
 
     const contentClassNames = cx([
@@ -247,4 +254,4 @@ PopupContent.Opts = {
   cornerSize: CORNER_SIZE,
 };
 
-export default PopupContent;
+export default injectIntl(PopupContent);