+++ packages/terra-modal-manager/CHANGELOG.md
@@ -4,6 +4,21 @@ Changelog
 Unreleased
 ----------
 
+5.0.0 - (January 25, 2019)
+------------------
+### Breaking Change
+* Updated terra-core dependencies
+
+4.3.0 - (January 5, 2019)
+------------------
+### Changed
+* Minor dependency update
+
+4.2.0 - (January 2, 2019)
+------------------
+### Changed
+* Updated dev site examples to not use forms or have duplicate ids
+
 4.1.0 - (December 5, 2018)
 ------------------
 ### Changed